requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
DateTime==4.3
Django==3.2.5
django-formtools==2.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Faker==8.10.1
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
zope.interface==5.4.0
uWSGI>=2.0.18,<2.1